Final integration get updates for deploy

+++ src/app/landing/components/list/house-item/house-item.component.ts
@@ -3,6 +3,8 @@ import {House} from '../../../models/house';
 import {WishlistService} from '../../../service/wishlist.service';
 import {DomSanitizer} from '@angular/platform-browser';
 import {Router} from '@angular/router';
+import {AuthenticationService} from '../../../../auth/services/authentication/authentication.service';
+import {MatSnackBar} from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-house-item',
@@ -12,7 +14,7 @@ import {Router} from '@angular/router';
 export class HouseItemComponent implements OnInit {
   // @Input() selected: boolean | undefined;
   // @Output() selectedChange = new EventEmitter<boolean>();
-  constructor(private wishlistService: WishlistService, public _sanitizer: DomSanitizer, private router: Router) { }
+  constructor( private snackbar: MatSnackBar,private wishlistService: WishlistService, public _sanitizer: DomSanitizer, private router: Router,private authService: AuthenticationService) { }
   @Input() item: House;
 
   @Input() addedToWishlist: boolean;
@@ -30,17 +32,31 @@ export class HouseItemComponent implements OnInit {
   }
 
   handleAddToWishlist() {
-    this.wishlistService.addToWishlist(this.item.houseID).subscribe((result) => {
-      console.log(result);
-      this.addedToWishlist = true;
-    });
+    if (this.authService.isAuthenticated()) {
+      this.wishlistService.addToWishlist(this.item.houseID).subscribe((result) => {
+        console.log(result);
+        this.addedToWishlist = true;
+      });
+    }
+    else{
+      this.snackbar.open('Please Login!', 'Close', {
+        duration: 3000
+      })
+    }
   }
 
   handleRemoveFromWishlist() {
-    this.wishlistService.removeFromWishlist(this.item.houseID).subscribe((result) => {
-      console.log(result);
-      this.addedToWishlist = false;
-    });
+    if (this.authService.isAuthenticated()) {
+      this.wishlistService.removeFromWishlist(this.item.houseID).subscribe((result) => {
+        console.log(result);
+        this.addedToWishlist = false;
+      });
+    }
+    else{
+      this.snackbar.open('Please Login!', 'Close', {
+        duration: 3000
+      })
+    }
   }
 }
